set up dependencies, file structure, views, & get requests for index.ejs
