+++ server.js
@@ -1,7 +1,3 @@
-// DEPENDENCIES
-const express = require('express');
-const app = express();
-
 // run `npm install` to install dependencies in package.json
 
 // * Your mission is to complete the app
@@ -19,63 +15,39 @@ const app = express();
 // views folder has not been created
 // views/missions folder has not been created
 
-// PORT
-const port = 3000;
-
+// DEPENDENCIES
+const express = require('express');
 // DATA - put into marsMissions.js file inside of a models folder, for module.exports
 // remember to require it in the server
-const marsMissions = [
-  {
-    name: "Curiosity",
-    launchDate: "26 Nov 2011",
-    operator: "NASA",
-    missionType: "Rover",
-    img: ""
-  },
-  {
-    name: "Opportunity",
-    launchDate: "8 Jul 2003",
-    operator: "NASA",
-    missionType: "Rover",
-    img: ""
-  },
-  {
-    name: "Spirit",
-    launchDate: "10 Jun 2003",
-    operator: "NASA",
-    missionType: "Rover",
-    img: ""
-  },
-  {
-    name: "Sojourner",
-    launchDate: "4 Dec 1996",
-    operator: "NASA",
-    missionType: "Rover",
-    img: ""
-  },
-  {
-    name: "Rosetta",
-    launchDate: "2 Mar 2004",
-    operator: "ESA",
-    missionType: "Gravity Assist",
-    img: ""
-  }
-];
+const marsMissions = require('./models/marsMissions.js')
+
+const app = express();
+
+// PORT
+const PORT = 3000;
+
+app.set('view engine', 'ejs') // this replaces (default template engine that comes with Express) with ejs
+
 
 // INDEX Route
 // send data to 'missions/index.ejs' view
 // the view should display just the names of each mission
 // display the mission names as <li> in a <ul> with the class name "missions"
-
+app.get('/', (request, response) => {
+response.render('missions/index.ejs', { marsMissions })
+})
 // SHOW Route
 // send data to 'missions/show.ejs' view
 // the view should display all the data for a single mission
+app.get('/:mission', (request, response) => {
+response.render('missions/show.ejs', { marsMissions })
+})
 
 
 
 // LISTENER
-app.listen(port, function() {
-  console.log('Missions to Mars running on port: ', port);
+app.listen(PORT, function() {
+  console.log('Missions to Mars running on port: ', PORT);
 })
 
 module.exports = app;
