+++ views/missions/show.ejs
@@ -1,0 +1,4 @@
+
+<% marsMissions.map( (mission, index) => { %>
+	<a href="#"><%= index.name %></a>
+<% }); %>
